Move player to corresponding module
