+++ app/src/main/AndroidManifest.xml
@@ -31,19 +31,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Krypt.Splash"
+        android:theme="@style/Theme.Krypt"
         tools:targetApi="31">
-        <activity
-            android:name=".ui.media.player.PlayerActivity"
-            android:exported="false"
-            android:launchMode="singleTop"
-            android:screenOrientation="sensor"
-            android:theme="@style/Theme.Krypt" />
 
         <activity
             android:name=".app.MainActivity"
             android:exported="true"
             android:launchMode="singleTask"
+            android:theme="@style/Theme.Krypt.Splash"
             android:windowSoftInputMode="stateHidden|adjustResize">
 
             <intent-filter>
