+++ app/src/main/res/values/strings.xml
@@ -36,8 +36,6 @@
     <string name="navigation_error">Error while navigation</string>
     <string name="password_confirmation">Password Confirm</string>
     <string name="grant">Grant</string>
-    <string name="cant_decrypt_video">Can\'t decrypt video</string>
-    <string name="decrypting_video_loading">Please wait until decrypt finished.</string>
     <string name="photos_folder_name">Photos</string>
     <string name="videos_folder_name">Videos</string>
     <string name="delete_file">Delete File</string>
