+++ features/media/player/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
     id("krypt.android.feature")
+    id("dagger.hilt.android.plugin")
 }
 
 android {
@@ -8,6 +9,10 @@ android {
 
 dependencies {
 
-//    implementation(libs.appCompat)
+    implementation(libs.androidx.appCompat)
+    implementation(libs.exoplayer)
+    implementation(libs.media3UI)
+    implementation(project(":shared:cryptography"))
+    implementation(project(":shared:files:files-logic"))
 
 }
