+++ features/media/player/src/main/java/ir/mehdiyari/krypt/mediaPlayer/PlayerViewModel.kt
@@ -1,4 +1,4 @@
-package ir.mehdiyari.krypt.ui.media.player
+package ir.mehdiyari.krypt.mediaPlayer
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
